menerapkan authorization

+++ src/api/notes/handler.js
@@ -16,8 +16,10 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload
-
-      const noteId = await this._service.addNote({ title, body, tags })
+      const { id: credentialId } = request.auth.credentials
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId
+      })
 
       const response = h.response({
         status: 'success',
@@ -48,7 +50,8 @@ class NotesHandler {
   }
 
   async getNotesHandler (request, h) {
-    const notes = await this._service.getNotes()
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
     return {
       status: 'success',
       data: {
@@ -60,7 +63,11 @@ class NotesHandler {
   async getNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       const note = await this._service.getNoteById(id)
+
       return {
         status: 'success',
         data: {
@@ -91,7 +98,8 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { id } = request.params
-
+      const { id: credentialId } = request.auth.credentials
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.editNoteById(id, request.payload)
 
       return {
@@ -122,17 +130,30 @@ class NotesHandler {
   async deleteNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.deleteNoteById(id)
       return {
         status: 'success',
         message: 'Catatan berhasil dihapus'
       }
     } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        })
+        response.code(error.statusCode)
+        return response
+      }
+
+      // Server ERROR!
       const response = h.response({
-        status: 'fail',
-        message: 'Catatan gagal dihapus. Id tidak ditemukan'
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.'
       })
-      response.code(404)
+      response.code(500)
+      console.error(error)
       return response
     }
   }
